+++ src/Engine/RendererBase.h
@@ -8,6 +8,8 @@ class RendererBase
 {
 public:
 	virtual void Initialize();
+	virtual void Cleanup();
+
 	virtual void PrecacheTexture(const std::string& name, const std::string& path);
 	virtual void RenderImage(const char* textureID, Vector2 position);
 	virtual void RenderImage(const char* textureID, Vector2 position, double scale, double rotation, SDL_RendererFlip flip);
@@ -15,14 +17,15 @@ public:
 	virtual void RenderFillRect(Vector2 position, Vector2 dimensions, SDL_Color color);
 	virtual void RenderPresent();
 	virtual void RenderClear();
+	virtual bool TextureExists(const std::string& Texture);
 
+	// Font Stuff
+	virtual void PrecacheFont(const std::string& name, const std::string& path, int fontsize);
+	virtual void RenderFont(const std::string& message, const char* fontID, Vector2 position, SDL_Color color);
+	virtual void RenderFont(const std::string& message, const char* fontID, Vector2 position, double rotation, SDL_RendererFlip flip, SDL_Color color);
 
 	virtual void SetRenderView(int x, int y);
 	virtual void SetRenderView(Vector2 pos);
-
-	virtual void Cleanup();
-	virtual bool TextureExists(const std::string& Texture);
-
 	virtual Vector2 VectorToRenderVector(Vector2 in);
 
 	virtual ~RendererBase();