+++ src/Engine/RendererSDL.cpp
@@ -11,19 +11,31 @@ void RendererSDL::Initialize()
 	int imgFlags = IMG_INIT_PNG | IMG_INIT_JPG;
 	if( !( IMG_Init( imgFlags ) & imgFlags ) )
 	{
-		printf( "SDL_image could not initialize! SDL_image Error: %s\n", SDL_GetError() ); return;
+		printf( "SDL_image could not initialize! SDL_image Error: %s\n", SDL_GetError() );
+		system("PAUSE");
+		exit(EXIT_FAILURE);
+	}
+
+	if (TTF_Init() < 0)
+	{
+		printf( "SDL_ttf could not initialize! SDL_ttf Error: %s\n", SDL_GetError() );
+		system("PAUSE");
+		exit(EXIT_FAILURE);
 	}
 
 	GameWindow = SDL_CreateWindow(Engine::Properties.Name.c_str(), SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED, Engine::Properties.Width, Engine::Properties.Height, SDL_WINDOW_SHOWN);
 	if(GameWindow == nullptr)
 	{
-		printf("Window could not be created! SDL_Error: %s\n", SDL_GetError()); return;
+		printf("Window could not be created! SDL_Error: %s\n", SDL_GetError());
+		system("PAUSE");
+		exit(EXIT_FAILURE);
 	}
 
 	GameRenderer = SDL_CreateRenderer(GameWindow, -1, SDL_RENDERER_ACCELERATED | SDL_RENDERER_PRESENTVSYNC);
 	if(GameRenderer == nullptr)
 	{
 		printf("Renderer could not be created! SDL_Error: %s\n", SDL_GetError());
+		system("PAUSE");
 		exit(EXIT_FAILURE);
 	}
 }
@@ -34,7 +46,7 @@ void RendererSDL::PrecacheTexture(const std::string& name, const std::string& pa
 	SDL_Surface* surface = IMG_Load(path.c_str());
 	if (surface == nullptr)
 	{
-		printf("Surface Load Fail: %s -> %s", name.c_str(), SDL_GetError());
+		printf("Texture precache Fail: %s -> %s", name.c_str(), SDL_GetError());
 		system("PAUSE");
 		exit(EXIT_FAILURE);
 	}
@@ -43,6 +55,7 @@ void RendererSDL::PrecacheTexture(const std::string& name, const std::string& pa
 	if(texture == nullptr)
 	{
 		printf("Texture could not be created! SDL_Error: %s\n", SDL_GetError());
+		system("PAUSE");
 		exit(EXIT_FAILURE);
 	}
 
@@ -103,6 +116,14 @@ void RendererSDL::RenderClear()
 }
 void RendererSDL::Cleanup()
 {
+	for(auto iter = FontRepository.begin(); iter != FontRepository.end(); iter++)
+	{
+		TTF_CloseFont(iter->second->Data);
+		iter->second->Data = nullptr;
+		delete iter->second;
+		iter->second = nullptr;
+	}
+
 	for(auto iter = TextureRepository.begin(); iter != TextureRepository.end(); iter++)
 	{
 		SDL_DestroyTexture(iter->second->Data);
@@ -113,11 +134,12 @@ void RendererSDL::Cleanup()
 	SDL_DestroyRenderer(GameRenderer);
 	SDL_DestroyWindow(GameWindow);
 	IMG_Quit();
+	TTF_Quit();
 }
 
 Vector2 RendererSDL::VectorToRenderVector(Vector2 in)
 {
-	//TODO: This is needless. -> Too Lazy to Remove.
+	//TODO: Improve this
 	// Displace coordinates.
 	in.x += RenderView.x;
 	in.y += RenderView.y;
@@ -127,4 +149,62 @@ Vector2 RendererSDL::VectorToRenderVector(Vector2 in)
 bool RendererSDL::TextureExists(const std::string& Texture)
 {
 	return (TextureRepository[Texture] != nullptr);
+}
+
+// SDL_ttf
+void RendererSDL::PrecacheFont(const std::string& name, const std::string& path, int fontsize)
+{
+	TTF_Font *font = TTF_OpenFont(path.c_str(), fontsize);
+	if(font == nullptr)
+	{
+		printf("Font precache fail: %s -> %s", name.c_str(), SDL_GetError());
+		system("PAUSE");
+		exit(EXIT_FAILURE);
+	}
+
+	FontRepository[name] = new SDLFontWrapper(font);
+	FontRepository[name]->Extras["Path"] = path;
+	FontRepository[name]->Extrai["Size"] = fontsize;
+}
+
+void RendererSDL::RenderFont(const std::string& message, const char* fontID, Vector2 position, double rotation, SDL_RendererFlip flip, SDL_Color color)
+{
+	std::string strID = std::string(fontID);
+	Vector2 pos = VectorToRenderVector(position);
+
+	if(FontRepository.count(strID) > 0
+	&& FontRepository[strID]->Data == nullptr)
+	{
+		printf("Font render attempt does not exist! -> %s\n", fontID);
+		return;
+	}
+
+	SDL_Surface *surf = TTF_RenderText_Blended(FontRepository[strID]->Data, message.c_str(), color);
+	if (surf == nullptr)
+	{
+		printf("Font surface render fail: %s -> %s", fontID, SDL_GetError());
+		system("PAUSE");
+		exit(EXIT_FAILURE);
+	}
+
+	SDL_Texture *texture = SDL_CreateTextureFromSurface(GameRenderer, surf);
+	if (texture == nullptr)
+	{
+		printf("Font texture from surface render fail: %s -> %s", fontID, SDL_GetError());
+		system("PAUSE");
+		exit(EXIT_FAILURE);
+	}
+
+
+	int texWidth, texHeight;
+	SDL_QueryTexture(texture, NULL, NULL, &texWidth, &texHeight);
+
+	SDL_Rect TexDimensions = {pos.x, pos.y, texWidth, texHeight};
+	SDL_RenderCopyEx(GameRenderer, texture, NULL, &TexDimensions, rotation, NULL, flip);
+
+}
+
+void RendererSDL::RenderFont(const std::string& message, const char* fontID, Vector2 position, SDL_Color color)
+{
+	this->RenderFont(message, fontID, position, 0, SDL_RendererFlip::SDL_FLIP_NONE, color);
 }